backend/requirements.txt
Flask==2.3.2
werkzeug==2.3.6

# Image Processing & Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-image==0.22.0
scipy==1.11.4

# 3D Processing
trimesh==4.0.5
shapely==2.0.2
open3d==0.17.0
pygltflib==1.16.1

# Machine Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200

# NLP & Speech
openai-whisper==20231117
spacy==3.7.2
pytesseract==0.3.10

# Database & Data Processing
pandas==2.1.3
sqlalchemy==2.0.23

# PDF & Reporting
reportlab==4.0.7

# API (Optional FastAPI support)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6